TinyLlama/requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
transformers
torch
faiss-cpu
